Update createAccountId method
